requirements.txt
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
attrs==23.1.0
boto3==1.34.2
botocore==1.34.2
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distro==1.8.0
einops==0.4.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.105.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.2
gradio==3.50.2
gradio_client==0.6.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-resources==6.1.1
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
more-itertools==10.1.0
mutagen==1.47.0
numba==0.58.1
numpy==1.26.2
openai==1.4.0
openai-whisper==20231117
opencv-python==4.5.5.64
orjson==3.9.10
packaging==23.2
pandas==2.1.4
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rpds-py==0.13.2
s3transfer==0.9.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
tiktoken==0.5.2
toolz==0.12.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.66.1
traitlets==5.14.0
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.24.0.post1
wcwidth==0.2.12
websockets==11.0.3
yt-dlp==2023.11.16
zipp==3.17.0
